+++ python/pathway/io/deltalake/__init__.py
@@ -59,9 +59,9 @@ def read(
     Reads a table from Delta Lake. Currently, local and S3 lakes are supported. The table
     doesn't have to be append only, however, the deletion vectors are not supported yet.
 
-    Note that the connector requires primary key fields to be specified in the schema.
-    You can specify the fields to be used in the primary key with ``pw.column_definition``
-    function.
+    Note that the connector requires either the table to be append-only or the primary key
+    fields to be specified in the schema. You can define the primary key fields using the
+    ``pw.column_definition`` function.
 
     Args:
         uri: URI of the Delta Lake source that must be read.
@@ -136,11 +136,6 @@ def read(
     are provided but the path starts with ``s3://`` or ``s3a://``, Pathway will use the
     credentials of the currently authenticated user.
     """
-    if schema.primary_key_columns() is None:
-        raise ValueError(
-            "DeltaLake reader requires explicit primary key fields specification"
-        )
-
     _check_entitlements("deltalake")
     prepared_connection_settings = _prepare_s3_connection_settings(
         s3_connection_settings